Refactor code into utils and tasks folders
