+++ vehicle/mkv/software/bms/BUILD
@@ -5,22 +5,45 @@ load("//bazel/tools:defs.bzl", "can_api_files")
 Runs on service section controller primary micro U1003, ATmega64m1
 """
 
-cc_firmware(
-    name = "bms",
+cc_library(
+    name = "bms_config",
     srcs = [
-        "bms.c",
+        "bms_config.c",
+    ],
+    hdrs = [
         "bms_config.h",
     ],
     target_compatible_with = [
         "//bazel/constraints:atmega64m1",
     ],
     deps = [
+        "//libs/adc",
+        "//libs/can",
+        "//libs/gpio",
+        "//libs/spi",
+        "//libs/timer",
+        "//vehicle/mkv/software/bms/ltc6811",
+    ],
+    visibility = ["//visibility:public"],
+)
+
+cc_firmware(
+    name = "bms",
+    srcs = [
+        "bms.c",
+    ],
+    target_compatible_with = [
+        "//bazel/constraints:atmega64m1",
+    ],
+    deps = [
+        ":bms_config",
         ":can_api",
         "//libs/adc",
         "//libs/gpio",
         "//libs/spi",
         "//libs/timer",
-        "//vehicle/mkv/software/bms/bsp",
+        "//vehicle/mkv/software/bms/utils",
+        "//vehicle/mkv/software/bms/tasks:bms_tasks",
         "//vehicle/mkv/software/bms/ltc6811",
     ],
 )
