+++ vehicle/mkv/software/bms/bms.c
@@ -9,17 +9,12 @@
 #include <avr/interrupt.h>
 #include <stdbool.h>
 
-#include "vehicle/mkv/software/bms/bsp/bsp.h"
+#include "vehicle/mkv/software/bms/utils/mux.h"
+#include "vehicle/mkv/software/bms/utils/can.h"
+#include "vehicle/mkv/software/bms/utils/fault.h"
 #include "vehicle/mkv/software/bms/can_api.h"
+#include "vehicle/mkv/software/bms/tasks/tasks.h"
 #include "vehicle/mkv/software/bms/ltc6811/ltc6811.h"
-
-/*
- * ic.system_open_wire is a 16 bit number. Each bit is an open wire (1: no
- * openwire, 0: openwire)
- */
-#define NO_OPEN_WIRES \
-    (65335) // ic.system_open_wire is a 16 bit number. Each
-            // bit is an open wire (1: no openwire, 0: openwire)
 
 // Defines maximum current range (-125 to 125A)
 #define MAX_CURRENT_RANGE (250)
@@ -36,20 +31,10 @@
 
 enum State {
     INIT = 0,
-    ACTIVE,
+    IDLE,
+    DISCHARGING,
     CHARGING,
     FAULT,
-};
-
-enum FaultCode {
-    FAULT_NONE = 0x00,
-    FAULT_UNDERVOLTAGE,
-    FAULT_OVERVOLTAGE,
-    FAULT_UNDERTEMPERATURE,
-    FAULT_OVERTEMPERATURE,
-    FAULT_DIAGNOSTICS_FAIL,
-    FAULT_OPEN_WIRE,
-    FAULT_OVERCURRENT,
 };
 
 // Represents all of LTC6811 chips used in the BMS. Contains cell
@@ -58,17 +43,6 @@ cell_asic ICS[NUM_ICS];
 
 uint16_t TEMPERATURES[NUM_TEMPERATURE_ICS][NUM_MUXES * NUM_MUX_CHANNELS];
 
-// Array of mux addresses
-const uint8_t MUXES[NUM_MUXES] = { LTC1380_MUX1, LTC1380_MUX2, LTC1380_MUX3 };
-
-static void set_fault(enum FaultCode the_fault) {
-    gpio_clear_pin(BMS_RELAY_LSD);
-
-    if (bms_core.fault_state == FAULT_NONE) {
-        bms_core.fault_state = the_fault;
-    }
-}
-
 /*
  * Interrupt for nOCD
  */
@@ -76,7 +50,7 @@ void pcint0_callback(void) {
     bms_core.overcurrent_detect = !gpio_get_pin(nOCD);
 
     if (bms_core.overcurrent_detect) {
-        set_fault(FAULT_OVERCURRENT);
+        set_fault(BMS_FAULT_OVERCURRENT);
     }
 }
 
@@ -85,13 +59,6 @@ void pcint0_callback(void) {
  */
 void pcint2_callback(void) {
     bms_core.bspd_current_sense = !!gpio_get_pin(BSPD_CURRENT_SENSE);
-}
-
-static void fan_enable(bool enable) {
-    timer1_fan_cfg.channel_b.pin_behavior = enable ? TOGGLE : DISCONNECTED;
-
-    // No way to update a single part of the config so we just re-init the timer
-    timer_init(&timer1_fan_cfg);
 }
 
 static uint16_t get_current_sensor_value(void) {
@@ -115,182 +82,6 @@ static void configure_ics(void) {
     LTC6811_wrcfg(NUM_ICS, ICS);
 }
 
-static void voltage_task(void) {
-    wakeup_sleep(NUM_ICS);
-
-    // Start cell voltage ADC conversions PLUS Sum of Cells Conversion
-    LTC6811_adcvsc(MD_7KHZ_3KHZ, DCP_ENABLED);
-
-    // Blocks until all ADCs are done being read
-    (void)LTC6811_pollAdc(); // Ignore return value because we don't care how
-                             // long it took
-
-    wakeup_idle(NUM_ICS);
-
-    // Stores the voltages in the c_codes variable
-    uint8_t error = LTC6811_rdcv(REG_ALL, NUM_ICS, ICS);
-
-    if (error) {
-        bms_metrics.voltage_pec_error_count += error;
-        return;
-    }
-
-    uint16_t pack_voltage = 0;
-    uint32_t uv = 0;
-    uint32_t ov = 0;
-
-    // Sum together all the voltages of the cells and check over/under voltage
-    // conditions
-    for (uint8_t ic = 0; ic < NUM_ICS; ic++) {
-        /*
-         * Sum of cells for a single IC
-         *
-         * The pins C5, C6, C11, and C12 are not connected to their own
-         * cells, so we subtract them off
-         */
-        pack_voltage += ICS[ic].stat.stat_codes[0]
-            - ICS[ic].cells.c_codes[4]
-            - ICS[ic].cells.c_codes[5]
-            - ICS[ic].cells.c_codes[10]
-            - ICS[ic].cells.c_codes[11];
-
-        // Odd bits of the ST register are used for undervoltage flags
-        uv = (ICS[ic].stat.flags[0] & 0x55) | (ICS[ic].stat.flags[1] & 0x55)
-             | (ICS[ic].stat.flags[2] & 0x55);
-
-        // Even bits of the ST register are used for overvoltage flags
-        ov = (ICS[ic].stat.flags[0] & 0xAA) | (ICS[ic].stat.flags[1] & 0xAA)
-             | (ICS[ic].stat.flags[2] & 0xAA);
-
-        if (uv > 0) {
-            set_fault(FAULT_UNDERVOLTAGE);
-        } else if (ov > 0) {
-            set_fault(FAULT_OVERVOLTAGE);
-        }
-    }
-
-    bms_core.pack_voltage = pack_voltage;
-}
-
-static void temperature_task(void) {
-    uint8_t error = 0;
-    int32_t cumulative_temperature = 0;
-
-    for (uint8_t mux = 0; mux < NUM_MUXES; mux++) {
-        // For each mux,
-        for (uint8_t ch = 0; ch < NUM_MUX_CHANNELS; ch++) {
-            // For each mux channel (ch)
-
-            enable_mux(NUM_ICS, ICS, MUXES[mux], MUX_ENABLE, ch);
-
-            // read aux gpio voltage (this is what the tmperature sensor is
-            // connected to) for temperature
-            LTC6811_adax(MD_7KHZ_3KHZ, AUX_CH_GPIO1);
-
-            // Wait for conversions to finish
-            (void)LTC6811_pollAdc();
-
-            // Read voltage from auxiliary pin (connected to the mux)
-            error = LTC6811_rdaux(AUX_CH_GPIO1, NUM_ICS, ICS);
-
-            if (error) {
-                bms_metrics.temperature_pec_error_count += error;
-                continue;
-            }
-
-            uint16_t temperature = 0;
-
-            /*
-             * temp_ic indexes the ICs that measure the temperature, so
-             *   temp_ic = 0 := ic 1
-             *   temp_ic = 1 := ic 3
-             *   temp_ic = 2 := ic 5
-             *   etc.
-             */
-            for (uint8_t temp_ic = 0; temp_ic < NUM_TEMPERATURE_ICS;
-                 temp_ic += 1) {
-                // Use 0 as the a_codes index for GPIO1 channel
-                temperature = ICS[(temp_ic * 2) + 1].aux.a_codes[0];
-
-                /*
-                 * Store temperature [(mux * NUM_MUX_CHANNELS) + ch] indexes the
-                 * channel of the mux. We store temperatures as a 2D array of
-                 * ICS vs mux channels and use 24 as the number of mux channels
-                 * (3 muxes, 8 channels each).
-                 */
-                uint16_t channel = (mux * NUM_MUX_CHANNELS) + ch;
-                TEMPERATURES[temp_ic][channel] = temperature;
-                cumulative_temperature += temperature;
-
-                /*
-                 * Using negative-temperature-coefficient (NTC) thermistors, so
-                 * comparisons are reversed (i.e. less than over-temp threshold)
-                 */
-                if (temperature < OVERTEMPERATURE_THRESHOLD) {
-                    set_fault(FAULT_OVERTEMPERATURE);
-                }
-
-                /*
-                 * TODO: It would be good to have some buffer so the fans aren't
-                 * pulsing on and off ever. This could be as simple as having a
-                 * counter while the fan is on, and we keep the fan on for the
-                 * duration of the counter, and once the counter is up, we check
-                 * if the temperature is in safe range (lower than soft OT), we
-                 * turn the fan off and keep checking.
-                 */
-
-                // If temperatures are getting a bit too high, we turn on the
-                // fan
-                if (temperature < SOFT_OVERTEMPERATURE_THRESHOLD) {
-                    fan_enable(true);
-                }
-
-                if (temperature > SOFT_OVERTEMPERATURE_THRESHOLD) {
-                    fan_enable(false);
-                }
-
-                if (temperature > UNDERTEMPERATURE_THRESHOLD) {
-                    set_fault(FAULT_UNDERTEMPERATURE);
-                }
-            } // End for each IC
-
-            /*
-             * Compute average temperature. WARNING: truncates a 32 bit number
-             * into a 16 bit number. This shouldn't be a problem, but the reader
-             * should be aware that this is intentional. A 32 bit number is
-             * needed to store the cumulative sum.
-             */
-            bms_core.temperature
-                = (uint16_t)cumulative_temperature
-                  / (NUM_MUXES * NUM_MUX_CHANNELS * NUM_TEMPERATURE_ICS);
-
-        } // End for each mux channel
-    } // End for each mux
-
-    // Finally, disable all multiplexers
-    enable_mux(NUM_ICS, ICS, MUX1, MUX_DISABLE, 0);
-    enable_mux(NUM_ICS, ICS, MUX2, MUX_DISABLE, 0);
-    enable_mux(NUM_ICS, ICS, MUX3, MUX_DISABLE, 0);
-}
-
-static void openwire_task(void) {
-    // Run open wire test on all cells
-    LTC6811_run_openwire_single(NUM_ICS, ICS);
-
-    long open_wire;
-    for (uint8_t ic = 0; ic < NUM_ICS; ic++) {
-        open_wire = ICS[ic].system_open_wire;
-
-        /*
-         * If system_open_wire is 0xFFFF, there are no open wires. Otherwise,
-         * the value is set to the first pin that has an open wire.
-         */
-        if (open_wire != NO_OPEN_WIRES) {
-            set_fault(FAULT_OPEN_WIRE);
-        }
-    }
-}
-
 /*
  * Perform initial startup checks on the BMS.
  *
@@ -304,7 +95,7 @@ static int initial_checks(void) {
 
     for (uint8_t ic = 0; ic < NUM_ICS; ic++) {
         if (ICS[ic].stat.mux_fail[0] == 1) {
-            set_fault(FAULT_DIAGNOSTICS_FAIL);
+            set_fault(BMS_FAULT_DIAGNOSTICS_FAIL);
             rc = 1;
             goto bail;
         }
@@ -319,7 +110,7 @@ static int initial_checks(void) {
     // read all voltages
     voltage_task();
 
-    if (bms_core.fault_state) {
+    if (bms_core.bms_fault_state) {
         rc = 1;
         goto bail;
     }
@@ -327,7 +118,7 @@ static int initial_checks(void) {
     // read all temperatures
     temperature_task();
 
-    if (bms_core.fault_state) {
+    if (bms_core.bms_fault_state) {
         rc = 1;
         goto bail;
     }
@@ -335,7 +126,7 @@ static int initial_checks(void) {
     // check for open-circuit
     openwire_task();
 
-    if (bms_core.fault_state) {
+    if (bms_core.bms_fault_state) {
         rc = 1;
         goto bail;
     }
@@ -391,7 +182,6 @@ int main(void) {
     gpio_clear_pin(GENERAL_LED);
 
     // Close the BMS shutdown circuit relay
-    // TODO: Unless charging
     gpio_set_pin(BMS_RELAY_LSD);
 
     sei();
@@ -401,6 +191,7 @@ int main(void) {
         uint8_t loop_counter = 0;
 
         if (run_10ms) {
+            // state_machine_run();
             wakeup_sleep(NUM_ICS);
 
             voltage_task();
