+++ vehicle/mkv/software/bms/tasks/tasks.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <stdint.h>
+
+/*
+ *
+ */
+void voltage_task(void);
+
+/*
+ *
+ */
+void temperature_task(void);
+
+/*
+ *
+ */
+void openwire_task(void);
