+++ vehicle/mkv/software/bms/utils/BUILD
@@ -1,23 +1,24 @@
 load("@rules_cc//cc:defs.bzl", "cc_library")
 
 cc_library(
-    name = "bsp",
+    name = "utils",
     srcs = [
-        "bsp.c",
         "can.c",
         "mux.c",
+        "fault.c",
     ],
     hdrs = [
-        "bsp.h",
         "can.h",
         "mux.h",
+        "fault.h",
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//libs/can",
         "//libs/gpio",
         "//libs/spi",
         "//libs/timer",
         "//vehicle/mkv/software/bms/ltc6811",
+        "//vehicle/mkv/software/bms:can_api",
+        "//vehicle/mkv/software/bms:bms_config",
     ],
 )
